requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.2
asgiref==3.8.1
blinker==1.9.0
boto3==1.37.33
botocore==1.37.33
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.2
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
fastavro==1.10.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
google-auth==2.38.0
griffe==1.7.2
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
logfire-api==3.14.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
mistralai==1.6.0
openai==1.73.0
opentelemetry-api==1.32.0
packaging==24.2
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-ai==0.0.55
pydantic-ai-slim==0.0.55
pydantic-evals==0.0.55
pydantic-graph==0.0.55
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0